nl2sql_server/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
pandas
tabulate
sqlparse
loguru
psycopg2-binary
python-dotenv
python-multipart
psycopg2
streamlit
pytest              
pytest-mock         
streamlit-testing
pytest-asyncio